Add files via upload
Simulation is able to generate random target formations as well as formations for surrounding and closing in on a target. Currently, one is able to surround a target with an integer valued radius (or  the one-norm deformation of a radius, since agents are confined to integer coordinates) and manually decrement that radius.